First attempt of porting on ESP8266/ESP32.
