+++ utils.cpp
@@ -1,6 +1,10 @@
  /*  the radian_trx SW shall not be distributed  nor used for commercial product*/
  /*  it is exposed just to demonstrate CC1101 capability to reader water meter indexes */
  /*  there is no Warranty on radian_trx SW */
+
+#include <arduino.h>
+
+#include "everblu_meters.h"
  
 void show_in_hex(uint8_t* buffer, size_t len)
 {
